Revert errorsnotesting build tag approach
